refactor: add EditorCore and Editor class for default ui
